[not verified] my-jetpack: add util for Product stories
