+++ projects/packages/my-jetpack/_inc/components/product-detail-card/stories/index.jsx
@@ -9,7 +9,7 @@ import withMock from 'storybook-addon-mock';
  * Internal dependencies
  */
 import ProductDetailCard, { ProductDetail } from '../index.jsx';
-import { backupProductData, boostProductData, scanProductData, searchProductData } from './mock-data.js';
+import { getMockData } from './utils.js';
 
 export default {
 	title: 'Packages/My Jetpack/Product Detail Card',
@@ -20,31 +20,6 @@ export default {
 const DefaultArgs = {};
 
 const DefaultProductDetailCard = args => <ProductDetailCard { ...args } slug="backup" />;
-
-const mapResponse = {
-	backup: backupProductData,
-	boost: boostProductData,
-	scan: scanProductData,
-	search: searchProductData,
-};
-
-/**
- * Helper function that returns the story mock data.
- *
- * @param {string} product - Product slug
- * @returns {Array}          Story mock data
- */
-export function getMockData( product ) {
-	const response = mapResponse[ product ];
-	return [
-		{
-			url: `my-jetpack/v1/site/products/${ product }?_locale=user`,
-			method: 'GET',
-			status: 200,
-			response,
-		},
-	];
-}
 
 export const _default = DefaultProductDetailCard.bind( {} );
 _default.parameters = {
